REB-17 App doesn't run migrations on startup (#15)
* fix: flyway migrations

* fix: application properties

+++ src/main/java/com/example/rebrain/services/CardService.java
@@ -18,7 +18,7 @@ public class CardService {
     private final CardRepo cardRepo;
 
     public CardEntity create(CardEntity cardEntity) {
-        Integer userId = ThreadLocalUserIdHolder.get();
+        Long userId = Long.valueOf(ThreadLocalUserIdHolder.get());
         cardEntity.setUserId(userId);
         log.debug("Saving new card: {}", cardEntity);
         CardEntity savedEntity = cardRepo.save(cardEntity);
